Refactored controller code into a common class used by both evo1 and evo2
