+++ control/__init__.py
@@ -1,0 +1,9 @@
+from control.evoland import evo_ctrl
+from control.menu_control import SeqLoadGame, SeqMenuConfirm, SeqMenuDown
+
+__all__ = [
+    "evo_ctrl",
+    "SeqLoadGame",
+    "SeqMenuDown",
+    "SeqMenuConfirm",
+]
