+++ evo2/combat/arena.py
@@ -1,7 +1,7 @@
 import logging
 import math
 
-import evo2.control
+from control import evo_ctrl
 from engine.mathlib import (
     Vec2,
     angle_between,
@@ -28,7 +28,7 @@ class SeqArenaCombat(SeqCombat):
         ]
 
     def _try_attack(self, target: GameEntity2D, weak_spot: Vec2) -> bool:
-        ctrl = evo2.control.handle()
+        ctrl = evo_ctrl()
         mem = get_zelda_memory()
         player_pos = mem.player.pos
         box = get_box_with_size(center=player_pos, half_size=self.precision)
